Fix readObject getting stuck.
Apparently, this happens because the server socket indefinitely awaits
on the line 'in.readObject();'. Perhaps this happens because the client
output stream isn't being flushed, but flushing at both ends didn't
solve the issue. Instead, what worked was creating a new socket
(therefore a new connection) every time a client tries to send a
command.

This means the ServerThreads have a short life span of executing just
one command.